Use inheritance for reuse benefits
